+++ CHANGELOG.md
@@ -19,6 +19,45 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.11] - 2026-03-18
+
+### Added
+
+- Add New component Modal with lili core v1
+- Add New component Dialog with lili core v1
+
+## [0.4.10] - 2026-03-18
+
+### Added
+
+- Add New component Alert with lili core v1
+- Add New component Accordion with lili core v1
+
+## [0.4.9] - 2026-03-15
+
+### Added
+
+- Add New component Card with lili core v1
+- Add New component AspectRatio with lili core v1
+
+## [0.4.8] - 2026-03-14
+
+### Added
+
+- Add New component Chip with lili core v1
+
+## [0.4.7] - 2026-03-12
+
+### Added
+
+- Add New component Separator with lili core v1
+
+## [0.4.6] - 2026-03-11
+
+### Added
+
+- Add New component Spacer with lili core v1
+
 ## [0.4.5] - 2026-03-09
 
 ### Added